Refactor controllers, add GitHub Actions workflow
Modified PersonController.cs to change the return type of the Put method to ActionResult<Person>? and removed commented-out code. Added an override for GetHashCode in Person.cs. Updated Program.cs to streamline CORS policy configuration and add Swagger generation services. Updated PersonRepository.cs to throw KeyNotFoundException if a person with the specified ID is not found. Added a new GitHub Actions workflow file, PromillePartnerBackEnd.yml, to automate build and deployment to Azure.